requirements.txt
attrs==25.4.0 ; python_version >= "3.10"
backports-datetime-fromisoformat==2.0.3 ; python_version == "3.10"
certifi==2026.1.4 ; python_version >= "3.10"
charset-normalizer==3.4.4 ; python_version >= "3.10"
click==8.3.1 ; python_version >= "3.10"
colorama==0.4.6 ; python_version >= "3.10" and platform_system == "Windows"
fsspec==2026.2.0 ; python_version >= "3.10"
greenlet==3.3.1 ; (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32") and python_version >= "3.10"
idna==3.11 ; python_version >= "3.10"
importlib-metadata==8.7.1 ; python_version < "3.12" and python_version >= "3.10"
inflection==0.5.1 ; python_version >= "3.10"
joblib==1.5.3 ; python_version >= "3.10"
jsonpath-ng==1.7.0 ; python_version >= "3.10"
jsonschema-specifications==2025.9.1 ; python_version >= "3.10"
jsonschema==4.26.0 ; python_version >= "3.10"
packaging==26.0 ; python_version >= "3.10"
pathlib-abc==0.5.2 ; python_version >= "3.10"
pendulum==3.2.0 ; python_version >= "3.10"
ply==3.11 ; python_version >= "3.10"
python-backoff==2.3.1 ; python_version >= "3.10"
python-dateutil==2.9.0.post0 ; python_version >= "3.10"
python-dotenv==1.2.1 ; python_version >= "3.10"
pyyaml==6.0.3 ; python_version >= "3.10"
referencing==0.37.0 ; python_version >= "3.10"
requests==2.32.5 ; python_version >= "3.10"
rpds-py==0.30.0 ; python_version >= "3.10"
simpleeval==1.0.3 ; python_version >= "3.10"
simplejson==3.20.2 ; python_version >= "3.10"
singer-sdk==0.53.5 ; python_version >= "3.10"
six==1.17.0 ; python_version >= "3.10"
sqlalchemy==2.0.46 ; python_version >= "3.10"
typing-extensions==4.15.0 ; python_version >= "3.10"
tzdata==2025.3 ; python_version >= "3.10"
universal-pathlib==0.3.9 ; python_version >= "3.10"
urllib3==2.6.3 ; python_version >= "3.10"
zipp==3.23.0 ; python_version < "3.12" and python_version >= "3.10"
